# src/main/resources/api/openapi-v1.yaml
openapi: '3.0.3'
info:
  title: MCRM Contact Service — Persons API
  version: '1.1.0'
  description: Person CRUD operations for MCRM contact-service

servers:
  - url: http://localhost:8081
    description: Local development

paths:
  /api/v1/persons:
    post:
      tags: [persons]
      summary: Создание Person
      operationId: createPerson
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePersonRequest'
      responses:
        '201':
          description: Person создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonResponse'
        '400':
          description: Ошибка валидации
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '409':
          description: Email уже существует
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
    get:
      tags: [ persons ]
      summary: Список Persons с фильтром по email и пагинацией
      operationId: listPersons
      parameters:
        - name: email
          in: query
          required: false
          schema:
            type: string
            format: email
          description: Точное совпадение email
        - name: page
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Page<PersonResponse>
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonPageResponse'
        '400':
          description: size > 100
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

  /api/v1/persons/{id}:
    get:
      tags: [ persons ]
      summary: Получение Person по ID
      operationId: getPerson
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Person найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonResponse'
        '404':
          description: Person не найден
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

    put:
      tags: [ persons ]
      summary: Обновление Person
      operationId: updatePerson
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePersonRequest'
      responses:
        '200':
          description: Person обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonResponse'
        '400':
          description: Ошибка валидации
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '404':
          description: Person не найден
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '409':
          description: Новый email уже занят
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

    delete:
      tags: [ persons ]
      summary: Удаление Person
      operationId: deletePerson
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Person удалён
        '404':
          description: Person не найден
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

components:
  schemas:
    CreatePersonRequest:
      type: object
      required:
        - fullName
        - email
      properties:
        fullName:
          type: string
          minLength: 1
          maxLength: 255
          example: "Иван Петров"
        email:
          type: string
          format: email
          maxLength: 255
          example: "ivan@example.com"
        phone:
          type: string
          maxLength: 50
          nullable: true
          example: "+79091234567"

    UpdatePersonRequest:
      type: object
      required:
        - fullName
        - email
      properties:
        fullName:
          type: string
          minLength: 1
          maxLength: 255
          example: "Иван Петров"
        email:
          type: string
          format: email
          maxLength: 255
          example: "ivan@example.com"
        phone:
          type: string
          maxLength: 50
          nullable: true
          example: "+79091234567"

    PersonResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        fullName:
          type: string
          example: "Иван Петров"
        email:
          type: string
          format: email
          example: "ivan@example.com"
        phone:
          type: string
          nullable: true
          example: "+79091234567"
        createdAt:
          type: string
          format: date-time
          example: "2026-05-05T13:30:00Z"
        updatedAt:
          type: string
          format: date-time
          nullable: true
          example: "2026-05-05T13:30:00Z"

    PersonPageResponse:
      type: object
      description: 'Page<PersonResponse>: items + page + size + total.'
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PersonResponse'
        page: { type: integer }
        size: { type: integer }
        total: { type: integer, format: int64 }

    ProblemDetail:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        service:
          type: string
        errorCode:
          type: string
        timestamp:
          type: string
          format: date-time
        traceId:
          type: string